intoreduce useGetStarted
